ai_modules organization refactor. Removes deprecates ai_client_service.py and tests.
