+++ backend/ai_modules/clients/__init__.py
@@ -1,0 +1,1 @@
+#!python3
